saving created user in DB
add 'create user service' & structuring schemas and resolvers using graphql-tools
