+++ package.json
@@ -13,6 +13,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@graphql-tools/load": "^8.0.0",
+    "@graphql-tools/load-files": "^7.0.0",
+    "@graphql-tools/merge": "^9.0.0",
     "apollo-server": "^3.12.1",
     "graphql": "^16.8.1",
     "pg": "^8.4.0",
